Move DownloadAppBadge into its own component
